buildtrace-overlay-/requirements.txt
opencv-python
numpy
matplotlib
scipy
pdf2image
pytesseract
Pillow
openai
python-dotenv
flask
flask-login
flask-dance
requests
werkzeug
gunicorn
PyMuPDF==1.26.4
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
google-cloud-storage==2.10.0
google-cloud-tasks==2.14.0
alembic==1.12.1
psutil
